fix: fixed generate first question error

+++ backend/app/features/selfaware/router.py
@@ -139,12 +139,15 @@ def create_or_get_today_question(
     question = question_service.get_questions_by_date(user.id, date)
     if not question:
         try: 
-            recent_question = question_service.list_questions_by_user(user.id, 1)[0]
-            question = question_service.generate_question(user.id)
-            # Delete unanswered question
-            recent_answer = answer_service.get_answer_by_question(recent_question.id)
-            if recent_answer == None:
-                question_service.delete_question_by_id(recent_question.id)
+            if question_service.list_questions_by_user(user.id, 1):
+                recent_question = question_service.list_questions_by_user(user.id, 1)[0]
+                question = question_service.generate_question(user.id)
+                # Delete unanswered question
+                recent_answer = answer_service.get_answer_by_question(recent_question.id)
+                if recent_answer == None:
+                    question_service.delete_question_by_id(recent_question.id)
+            else:
+                question = question_service.generate_question(user.id)
         except IntegrityError:
             question = question_service.get_questions_by_date(user.id, date)
         assert type(question) == Question
